refactor(tests): consolidate insight tests following PR #420 pattern
Move getInstanceStatusByDate tests into existing cases.test.ts and
processes.test.ts instead of separate files. Add shared integration
test helper testGetInstanceStatusByDate() to avoid duplication.
