+++ tests/integration/shared/maestro/cases.integration.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { getServices, setupUnifiedTests, InitMode } from '../../config/unified-setup';
-import { testGetTopRunCount } from '../../utils/helpers';
+import { testGetTopRunCount, testGetInstanceStatusByDate } from '../../utils/helpers';
 
 const modes: InitMode[] = ['v0', 'v1'];
 
@@ -44,39 +44,7 @@ describe.each(modes)('Maestro Cases - Integration Tests [%s]', (mode) => {
   describe('getInstanceStatusByDate', () => {
     it('should retrieve instance status by date for case management', async () => {
       const { cases } = getServices();
-
-      const now = Date.now();
-      const sevenDaysAgo = now - 7 * 24 * 60 * 60 * 1000;
-
-      const result = await cases.getInstanceStatusByDate(sevenDaysAgo, now);
-
-      expect(result).toBeDefined();
-      expect(Array.isArray(result)).toBe(true);
-
-      if (result.length > 0) {
-        const entry = result[0];
-        expect(entry.startTime).toBeDefined();
-        expect(typeof entry.startTime).toBe('string');
-        expect(entry.status).toBeDefined();
-        expect(typeof entry.status).toBe('string');
-        expect(entry.count).toBeDefined();
-        expect(typeof entry.count).toBe('number');
-      }
-    });
-
-    it('should support timeSliceUnit option', async () => {
-      const { cases } = getServices();
-      const { TimeSliceUnit } = await import('../../../../src/models/common');
-
-      const now = Date.now();
-      const sevenDaysAgo = now - 7 * 24 * 60 * 60 * 1000;
-
-      const result = await cases.getInstanceStatusByDate(sevenDaysAgo, now, {
-        timeSliceUnit: TimeSliceUnit.Hour,
-      });
-
-      expect(result).toBeDefined();
-      expect(Array.isArray(result)).toBe(true);
+      await testGetInstanceStatusByDate(cases);
     });
   });
